engine/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
requests>=2.32.0
loguru>=0.7.2
python-dotenv>=1.0.0
